use prop-types, use cherry pick css antd
