+++ README.md
@@ -43,9 +43,9 @@ const App = () => {
           setValue(e);
         }}
         color={"red"}
-        maxWidth={'250px'}
+        maxWidth={"250px"}
         width={"100%"}
-        height={'35px'}
+        height={"35px"}
         marginIMG={"0px 10px"}
         marginButton={"10px auto"}
         // autoFocus={true}
@@ -72,12 +72,34 @@ const App = () => {
         step={2}
       />
 
+      Step 2 disable
+      <CLFInputNumber
+        imgComponent={<IdentityImage/>}
+        value={7}
+        step={2}
+        disabled={true}
+      />
+
+      Step 2 with no img disable
+      <CLFInputNumber
+        value={7}
+        step={2}
+        disabled={true}
+      />
+
+      Step 2 with no img readOnly
+      <CLFInputNumber
+        value={7}
+        step={2}
+        readOnly={true}
+      />
+
+
     </div>
   );
-}
+};
 
 export default App;
-
 
 ```
