+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "input-numbers",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "description": "ClassFunc library input number component use easy!",
   "author": "duongdam",
   "license": "MIT",
@@ -44,6 +44,7 @@
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
+    "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-scripts": "^3.4.1"
